fix(plugin-sdk): keep reply payload entrypoint lightweight
